services/auth/requirements.txt
fastapi
uvicorn[standard]
pydantic
prometheus-fastapi-instrumentator
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
asyncpg
sqlalchemy[asyncio]
python-multipart

